Models tab and add model dialog now work with the modelManagementController, previews are not yet implemented

+++ Controllers/Converter/ColorConverter.cs
@@ -17,9 +17,9 @@ namespace Controllers.Converter
         {
             ColorDTO colorDTO = new ColorDTO()
             {
-                Red = (int)color.Red,
-                Green = (int)color.Green,
-                Blue = (int)color.Blue,
+                Red = color.Red,
+                Green = color.Green,
+                Blue = color.Blue,
             };
 
             return colorDTO;
